Improve Facebook error handling
If the Facebook authenticator fails, then it does not send a `code`
request to our callback page. Since this is the result of a legitimate
human interaction (e.g. declining to grant SMR permissions), we do not
want to raise an Exception, which was currently happening due to the
`Request::get('code')` call. Instead, we now send the user back to the
login page with a helpful error message.

This is intended to address a warning from Facebook Operations that
would cause our app to have its approval revoked:

> Developer Policy 6.1: Verify that you have integrated Login correctly.
> Your app shouldn't crash or hang during the testing process.